requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.15 